// web-next/src/api/schema.ts
export interface components {
    schemas: {
// ...
        ApiResponseWebMe: {
            data: components["schemas"]["WebMe"] | null;
            error: components["schemas"]["ErrorDetails"] | null;
            success: boolean;
        };
// ...
        ErrorDetails: {
            /** @enum {string} */
            code: "invalid_request" | "invalid_request_data" | "validation_error" | "unauthorized" | "permission_denied" | "access_denied" | "restricted" | "csrf" | "not_found" | "business_rule_violation" | "conflict" | "rate_limited" | "internal_server_error";
            message: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}